==> Experiments/lsst_forecast/step00_ground_truth_kin.ipynb <==
# Step 0: Compute ground truth kinematics and add to catalog #

### Note: This notebook runs in the forecast_env ###

This is quite expensive, we'd really only like to do this once. Let's pre-compute and track it as we go...

%load_ext autoreload
%autoreload 2
import pandas as pd
import sys
from astropy.cosmology import FlatLambdaCDM
sys.path.insert(0, '/Users/smericks/Desktop/StrongLensing/darkenergy-from-LAGN/')
import Utils.ground_truth_utils as gt_utils

# ground truth cosmology
gt_cosmo = FlatLambdaCDM(H0=70.,Om0=0.3)

# take in om10 catalog
lenses = pd.read_csv('DataVectors/om10_venkatraman25.csv')

already_run = True # takes ~23mins to run, hence this flag 

if not already_run: 
    # add a column with kinematics
    gt_utils.populate_truth_sigma_v_4MOST(lenses,gt_cosmo)

    # re-save
    lenses.to_csv('DataVectors/om10_venkatraman25_withkin.csv',index=False)
